move market utils into lend namespace
